Shorten and simplify code
I made it so when you use 'open', you don't have to close it by using "with open() as" instead. I also added some uses of the walrus operator in order to shorten your code by setting variables inside print statements and other function statements.
Note: If this is not intended for use with python 3.8 or above, please discard this pull request as the walrus operator will not be recognized on earlier versions.